requirements.txt
absl_py==2.1.0
isaacgym.egg==info
latextable==1.0.1
matplotlib==3.7.5
natsort==8.4.0
numpy==1.23.5
opencv_python==4.9.0.80
Pillow==10.2.0
psutil==5.9.8
scipy==1.12.0
setuptools==59.5.0
tensorboard==2.14.0
texttable==1.7.0
torch==1.10.0+cu113
wandb==0.16.3
